src/requirements.txt
autopep8
pycodestyle
pypresence
PySide6
python-dotenv
shiboken6
toml
pathlib
webbrowser
subprocess
platform
logging
pyqtdarktheme
zlib